phosophere plotter file
